improve certificate selection on macOS

+++ src/MQTTAnalyzer/icloud/CloudDataManager.swift
@@ -18,7 +18,7 @@ class CloudDataManager {
     struct DocumentsDirectory {
         static let localDocumentsURL = FileManager.default
 			.urls(for: FileManager.SearchPathDirectory.documentDirectory, in: .userDomainMask)
-			.last!
+			.last
 		
         static let iCloudDocumentsURL = FileManager.default
 			.url(forUbiquityContainerIdentifier: nil)?
@@ -72,22 +72,27 @@ class CloudDataManager {
 		return DocumentsDirectory.iCloudDocumentsURL
     }
 
-	func getLocalDocumentDiretoryURL() -> URL {
+	func getLocalDocumentDiretoryURL() -> URL? {
 		return DocumentsDirectory.localDocumentsURL
     }
 	
 	func copyFileToLocal(file: URL) {
 		let fileManager = FileManager.default
 		do {
-			let target = DocumentsDirectory.localDocumentsURL.appendingPathComponent(file.lastPathComponent)
-			
-			if fileManager.fileExists(atPath: target.path) {
-				try fileManager.removeItem(at: target)
+			if let url = DocumentsDirectory.localDocumentsURL {
+				let target = url.appendingPathComponent(file.lastPathComponent)
+				
+				if fileManager.fileExists(atPath: target.path) {
+					try fileManager.removeItem(at: target)
+				}
+				
+				try fileManager.copyItem(at: file,
+										 to: url.appendingPathComponent(file.lastPathComponent))
+				CloudDataManager.logger.debug("Copied \(file) to local dir")
 			}
-			
-			try fileManager.copyItem(at: file,
-									 to: DocumentsDirectory.localDocumentsURL.appendingPathComponent(file.lastPathComponent))
-			CloudDataManager.logger.debug("Copied \(file) to local dir")
+			else {
+				CloudDataManager.logger.error("Local URL not found")
+			}
 		} catch let error as NSError {
 			CloudDataManager.logger.error("Failed to copy file to local directory: \(error)")
 		}
@@ -96,8 +101,14 @@ class CloudDataManager {
 	func deleteLocalFile(fileName: String) {
 		let fileManager = FileManager.default
 		do {
-			try fileManager.removeItem(at: DocumentsDirectory.localDocumentsURL.appendingPathComponent(fileName))
-			CloudDataManager.logger.debug("Deleted file \(fileName)")
+			if let url = DocumentsDirectory.localDocumentsURL {
+				try fileManager.removeItem(at: url.appendingPathComponent(fileName))
+				CloudDataManager.logger.debug("Deleted file \(fileName)")
+			}
+			else {
+				CloudDataManager.logger.error("Local URL not found")
+			}
+			
 		} catch let error as NSError {
 			CloudDataManager.logger.error("Failed to delete file: \(error)")
 		}
@@ -112,11 +123,17 @@ extension CertificateFile {
 			}
 			else {
 				CloudDataManager.logger.error("No cloud URL found (Cloud disbled?)")
-				throw CertificateError.noClound
+				throw CertificateError.noCloud
 			}
 		}
 		else {
-			return CloudDataManager.instance.getLocalDocumentDiretoryURL()
+			if let url = CloudDataManager.instance.getLocalDocumentDiretoryURL() {
+				return url
+			}
+			else {
+				CloudDataManager.logger.error("No local URL found")
+				throw CertificateError.noLocalURL
+			}
 		}
 	}
 }
